requirements.txt
# Cannabis Data Aggregator - Python Dependencies
# Install with: pip install --prefer-binary -r requirements.txt

# Web Framework
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.3

# Database
SQLAlchemy==2.0.31
alembic==1.13.2
PyMySQL==1.1.1             # MySQL support
cryptography==43.0.3       # Required by PyMySQL auth

# Scheduling
APScheduler==3.10.4

# HTTP / Data Fetching
requests==2.32.3
httpx==0.27.0

# Data Processing (install separately: pip install --only-binary :all: pandas numpy)
# pandas and numpy require compiled wheels — see install instructions
openpyxl==3.1.5            # Excel support

# GeoJSON / Geospatial
geojson==3.1.0

# Config / Environment
python-dotenv==1.0.1
PyYAML==6.0.2

# Data Validation
pydantic==2.8.2

# Utilities
python-dateutil==2.9.0
pytz==2024.1
chardet==5.2.0             # Encoding detection
click==8.1.7
rich==13.7.1               # Pretty terminal output
tqdm==4.66.5               # Progress bars

# Caching
cachetools==5.5.0

# Optional: PostgreSQL support (uncomment if using PostgreSQL)
# psycopg2-binary==2.9.9
